// src/app/store/actions.types.ts
export enum ActionType{
    MENU_START = "[MENU] MENU START",
    MENU_SUCCESS = "[MENU] MENU SUCCESS",
    MENU_ERROR = "[MENU] MENU ERROR",
    FLOORTYPE_START = "[FLOOR] FLOOR START",
    FLOORTYPE_SUCCESS = "[FLOOR] FLOOR SUCCESS",
    FLOORTYPE_ERROR = "[FLOOR] FLOOR ERROR",
    SALETITLEHEAD_START = "[SALE] SALE HEAD START",
    SALETITLEHEAD_SUCCESS = "[SALE] SALE HEAD SUCCESS",
    SALETITLEHEAD_ERROR = "[SALE] SALE HEAD ERROR",
    
    SALETITLEFOOTER_START = "[SALE-FOOTER] SALE FOOTER START",
    SALETITLEFOOTER_SUCCESS = "[SALE-FOOTER] SALE FOOTER SUCCESS",
    SALETITLEFOOTER_ERROR = "[SALE-FOOTER] SALE FOOTER ERROR",
}